Support decimal inventory quantities and align POS docs
- Add migration and model/service updates for decimal stock on purchases and ledger

- Extend Purchases API and api-spec; add/adjust feature tests

- Receipt templates, X-report modal, and FlutterFlow DSL/docs/custom actions

+++ app/Services/PurchaseService.php
@@ -1482,9 +1482,9 @@ class PurchaseService
                     $itemId = $refundedItem['item_id'] ?? null;
                     if ($itemId) {
                         if (! isset($itemRefunds[$itemId])) {
-                            $itemRefunds[$itemId] = 0;
+                            $itemRefunds[$itemId] = 0.0;
                         }
-                        $itemRefunds[$itemId] += $refundedItem['quantity'] ?? 1;
+                        $itemRefunds[$itemId] += (float) ($refundedItem['quantity'] ?? 1);
                     }
                 }
                 $metadata['item_refunds'] = $itemRefunds;
